Replace old input system code with new
Jump is set to listen to both press and release, therefore, an extra check is needed to make it only jump on press. We can change the interaction if we don't want that, but it'd be handy for increasing gravity when the button is released or similar, if we wanted to do that.

The camera controller will need its look speed to be reduced. I think it feels less jittery when it requires EnablePointerLook to be down, which is weird.