Fix camera setup, mostly "perfect" collision
Ideally, the camera would move forward up until minimum distance, even if that position would be clipping inside the obstacle. Even that might cause weirdness, though.

If such a thing isn't possible (and perhaps even if it is), we'll need a method of keeping the player visible through objects for when the camera ignores a collision due to it being too close.