Added temp button image + demo stuff
"Demo stuff" is a few triggers in the CinemachineTest scene to play around with.

The "temp" button image is made out of inbuilt Unity sprites and TextMeshPro, but it honestly looks good as is, so unless we want to do something more fancy to indicate the buttons, this may as well not be temp.