Fix previous commit errors, move audio listener
Now there's an audio listener in the player that rotates with the camera, so we hear the sounds from the player's position, not the orbiting camera.

The camera still has an audio listener, but it's disabled. If we add cutscenes where the camera detaches from the player in the future, we can enable it and disable the player's.