grab init script assigns mesh renderer to sparkles
This means the mesh will be set up to work in builds automatically by unity; Unity only knows it should do that if the mesh is referenced somewhere like a particle system in a prefab/scene. It doesn't know it should do that if it's set dynamically via script, like this code used to do.