Hide dart board, "x10" instead of "x10-14"
The dartboard distracts from the more important mote requirement on the door. (Why does a kid this age have a dart board anyway?)

I know the "-" is a hyphen, therefore, 10 _to_ 14, but it could also be read as 10 _minus_ 14, and generally takes up more space than it needs to. This eliminates that potential confusion.